Fix compiler warnings

+++ src/axes.c
@@ -530,10 +530,10 @@ void vky_axes_make_vertices(
     const int32_t NxC = (NXC - 1) * ((int32_t)MIT - 1);
     const int32_t NyC = (NYC - 1) * ((int32_t)MIT - 1);
 
-    double* TX = calloc(NXC, sizeof(double));
-    double* TY = calloc(NYC, sizeof(double));
-    double* Tx = calloc(NxC, sizeof(double));
-    double* Ty = calloc(NyC, sizeof(double));
+    double* TX = calloc((size_t)NXC, sizeof(double));
+    double* TY = calloc((size_t)NYC, sizeof(double));
+    double* Tx = calloc((size_t)NxC, sizeof(double));
+    double* Ty = calloc((size_t)NyC, sizeof(double));
 
     // Minor ticks, major ticks, grid, lims, user.
     const int64_t N = 2 * NXC + 2 * NYC + NxC + NyC + 2 + (int32_t)axes->user.tick_count;
